+++ include/core/Stream.h
@@ -69,7 +69,7 @@ namespace jm
           \details This method returns the length of the stream in bytes.
           \return The length of the stream.
           */
-         virtual int64 size() const = 0;
+         virtual size_t size() const = 0;
 
          /*!
           \brief Opens the stream for read or write operations.
@@ -104,7 +104,7 @@ namespace jm
           If you want to ensure that all data is read, you should call readFully().
           \return The actual number of bytes read, or 0 if no bytes were read (EOF).
           */
-         virtual int64 read(uint8* buffer, int64 length) = 0;
+         virtual size_t read(uint8* buffer, size_t length) = 0;
 
          /*!
           \brief Reads a maximum of length bytes into the array.
@@ -115,14 +115,14 @@ namespace jm
            \param length The maximum number of bytes to read.
           \return The actual number of bytes read, or 0 if no bytes were read (EOF).
           */
-         virtual int64 readFully(ByteArray& buffer, int64 length) = 0;
+         virtual size_t readFully(ByteArray& buffer, size_t length) = 0;
 
          /*!
          \brief Reads a maximum of length bytes into the array.
          \param buffer The buffer to read the data from.
          \return The actual number of bytes read, or 0 if no bytes were read (EOF).
          */
-         int64 readFully(ByteArray& buffer);
+         size_t readFully(ByteArray& buffer);
 
          /*!
           \brief Moves the file cursor to the desired position, counted from the beginning of the file
@@ -130,13 +130,13 @@ namespace jm
           \note Not every stream supports this method.
           \details This method moves the file cursor to the specified position, counted from the beginning of the file.
           */
-         virtual void seek(int64 position) = 0;
+         virtual void seek(size_t position) = 0;
 
          /*!
           \brief Moves the file cursor to the desired position, counted from the current position.
           \details This method moves the file cursor to the specified position, counted from the current position.
           */
-         virtual void move(int64 offset) = 0;
+         virtual void move(ssize_t offset) = 0;
 
          /*!
           \brief Returns the current cursor position in the file.
@@ -144,7 +144,7 @@ namespace jm
           \details This method returns the current cursor position in the file.
           \return The current cursor position.
           */
-         virtual int64 position() = 0;
+         virtual size_t position() = 0;
 
          /*!
           \brief Writes a buffer to the output file.
@@ -153,7 +153,7 @@ namespace jm
           \param length The number of bytes to write from the buffer.
           \return The actual number of bytes written.
           */
-         virtual int64 write(const uint8* buffer, int64 length) = 0;
+         virtual size_t write(const uint8* buffer, size_t length) = 0;
 
          /*!
           \brief Writes a buffer to the output file.
@@ -162,13 +162,13 @@ namespace jm
           \param length The number of bytes to write from the buffer.
           \return The actual number of bytes written.
           */
-         int64 write(const int8* buffer, int64 length);
+         size_t write(const int8* buffer, size_t length);
 
          /*!
           \brief Writes a string to the output file. The encoding is the default encoding (UTF-8). NOT AS A C-STRING!
           \details This method writes the specified string to the output file using the default encoding, which is UTF-8. The string should not be passed as a C-string.
           */
-         int64 write(const String& string);
+         size_t write(const String& string);
 
    };
 }
